Revert "리덕스 리팩토링, 캘린더 컴포넌트 관련 PR 입니다. (#8)"
This reverts commit fb706385c7b5ae597ccf42ea31e6dfb2b0fb3000.
