+++ package.json
@@ -10,7 +10,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@nivo/calendar": "^0.87.0",
     "@reduxjs/toolkit": "^2.2.5",
     "@supabase/supabase-js": "^2.43.4",
     "react": "^18.2.0",
