+++ src/Redux/Config/configureStore.js
@@ -1,11 +1,9 @@
 import { configureStore } from '@reduxjs/toolkit';
 import postModalReducer from '../Slices/PostModalSlice';
-import postAddReducer from '../Slices/PostsSlice';
 
 const store = configureStore({
   reducer: {
-    modal: postModalReducer,
-    posts: postAddReducer
+    modal: postModalReducer
   }
 });
