+++ src/components/MyPage/PostingList.jsx
@@ -1,16 +1,30 @@
+import React, { useEffect, useState } from 'react';
+import { useAuth } from '../../contexts/Auth/hooks';
 import styled from 'styled-components';
-import UseFetchPosts from './hooks/UseFetchPosts';
-import { useSelector } from 'react-redux';
 
 export function PostingList() {
-  const posts = useSelector((state) => state.posts.posts);
+  const { supabaseClient, user } = useAuth();
+  const [postData, setPostData] = useState([]);
 
-  UseFetchPosts(); // 커스텀 훅 호출
+  // 마운트 될때 posts 데이터 가져오기
+  useEffect(() => {
+    const fetchPostsData = async () => {
+      if (user) {
+        const { data: fetchPostsData } = await supabaseClient
+          .from('posts')
+          .select('id, UID, title, content, created_at, img_url')
+          .eq('UID', user.user_metadata.userName);
+
+        setPostData(fetchPostsData);
+      }
+    };
+    fetchPostsData();
+  }, [user, supabaseClient]);
 
   return (
     <>
       <StListWrap>
-        {posts.map((post) => (
+        {postData.map((post) => (
           <StPostItem key={post.id}>
             <h3>{post.title}</h3>
             <p>{post.content}</p>
